Made address book work with individual commands
Changed the code and made it more neat to segregate it into "people" and "wallet" commands
