+++ src/main/java/seedu/address/logic/commands/people/FindCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.people;
 
 import static java.util.Objects.requireNonNull;
