+++ src/main/java/seedu/address/logic/parser/Parser.java
@@ -1,6 +1,6 @@
 package seedu.address.logic.parser;
 
-import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.people.Command;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
