+++ src/main/java/seedu/address/logic/parser/SharkieParser.java
@@ -1,0 +1,104 @@
+package seedu.address.logic.parser;
+
+import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+import seedu.address.logic.commands.people.AddCommand;
+import seedu.address.logic.commands.people.ClearCommand;
+import seedu.address.logic.commands.people.Command;
+import seedu.address.logic.commands.people.DeleteCommand;
+import seedu.address.logic.commands.people.EditCommand;
+import seedu.address.logic.commands.people.ExitCommand;
+import seedu.address.logic.commands.people.FindCommand;
+import seedu.address.logic.commands.people.HelpCommand;
+import seedu.address.logic.commands.people.ListCommand;
+import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.logic.parser.people.AddCommandParser;
+import seedu.address.logic.parser.people.DeleteCommandParser;
+import seedu.address.logic.parser.people.EditCommandParser;
+import seedu.address.logic.parser.people.FindCommandParser;
+
+/**
+ * Parses user input.
+ */
+public class SharkieParser {
+
+    /**
+     * Used for initial separation of command word and args.
+     */
+    private static final Pattern BASIC_COMMAND_FORMAT = Pattern.compile("(?<commandType>\\S+) (?<commandWord>\\S+)(?<arguments>.*)");
+    private static final String PEOPLE_COMMAND_TYPE = "people";
+    private static final String WALLET_COMMAND_TYPE = "wallet";
+
+    /**
+     * Parses user input into command for execution.
+     *
+     * @param userInput full user input string
+     * @return the command based on the user input
+     * @throws ParseException if the user input does not conform the expected format
+     */
+    public Command parseCommand(String userInput) throws ParseException {
+        final Matcher matcher = BASIC_COMMAND_FORMAT.matcher(userInput.trim());
+        if (!matcher.matches()) {
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, HelpCommand.MESSAGE_USAGE));
+        }
+
+        final String commandType = matcher.group("commandType");
+        final String commandWord = matcher.group("commandWord");
+        final String arguments = matcher.group("arguments");
+
+        switch(commandType) {
+        case PEOPLE_COMMAND_TYPE:
+            return getPeopleCommand(commandWord, arguments);
+        case WALLET_COMMAND_TYPE:
+            return getWalletCommand(commandWord, arguments);
+        default:
+            throw new ParseException(MESSAGE_UNKNOWN_COMMAND);
+        }
+
+    }
+
+    private Command getPeopleCommand(String commandWord, String arguments) throws ParseException {
+        switch (commandWord) {
+
+            case AddCommand.COMMAND_WORD:
+                return new AddCommandParser().parse(arguments);
+
+            case EditCommand.COMMAND_WORD:
+                return new EditCommandParser().parse(arguments);
+
+            case DeleteCommand.COMMAND_WORD:
+                return new DeleteCommandParser().parse(arguments);
+
+            case ClearCommand.COMMAND_WORD:
+                return new ClearCommand();
+
+            case FindCommand.COMMAND_WORD:
+                return new FindCommandParser().parse(arguments);
+
+            case ListCommand.COMMAND_WORD:
+                return new ListCommand();
+
+            case ExitCommand.COMMAND_WORD:
+                return new ExitCommand();
+
+            case HelpCommand.COMMAND_WORD:
+                return new HelpCommand();
+
+            default:
+                throw new ParseException(MESSAGE_UNKNOWN_COMMAND);
+        }
+    }
+
+    private Command getWalletCommand(String commandWord, String arguments) throws ParseException {
+        switch(commandWord) {
+
+        default:
+            throw new ParseException(MESSAGE_UNKNOWN_COMMAND);
+        }
+    }
+
+}
