+++ src/test/java/seedu/address/logic/commands/CommandResultTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
+import seedu.address.logic.commands.people.CommandResult;
 
 public class CommandResultTest {
     @Test
