+++ src/test/java/seedu/address/logic/commands/EditCommandTest.java
@@ -18,7 +18,9 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.EditCommand.EditPersonDescriptor;
+import seedu.address.logic.commands.people.ClearCommand;
+import seedu.address.logic.commands.people.EditCommand;
+import seedu.address.logic.commands.people.EditCommand.EditPersonDescriptor;
 import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
